move algorithms for indivdidual scores to solidago
