+++ solidago/src/solidago/mehestan/individual/__init__.py
@@ -1,0 +1,2 @@
+from .indiv_matrix_inversion import MatrixInversionInvididualAlgo
+from .continuous_bradly_terry import ContinuousBradleyTerry
